+++ tests/CircleTest.php
@@ -28,7 +28,7 @@ use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgAttri
  * Test coverage.
  * - Accurate rendering of the `<circle>` element.
  * - Correct application of global HTML attributes and SVG-specific attributes.
- * - Error handling for invalid attributes or configuration.
+ * - Error handling for invalid attribute values.
  * - Immutability of the API, ensuring that setting attributes returns a new instance.
  * - Integration with configuration providers and global factory defaults.
  * - Precedence of user-defined attributes over global defaults and provider settings.
@@ -538,7 +538,7 @@ final class CircleTest extends TestCase
         );
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidFillOpacityValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingFillOpacityValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -548,7 +548,7 @@ final class CircleTest extends TestCase
         Circle::tag()->fillOpacity('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidFillRuleValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingFillRuleValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -562,7 +562,7 @@ final class CircleTest extends TestCase
         Circle::tag()->fillRule('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidOpacityValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingOpacityValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -572,7 +572,7 @@ final class CircleTest extends TestCase
         Circle::tag()->opacity('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineCapValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingStrokeLineCapValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -586,7 +586,7 @@ final class CircleTest extends TestCase
         Circle::tag()->strokeLineCap('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeLineJoinValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingStrokeLineJoinValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -600,7 +600,7 @@ final class CircleTest extends TestCase
         Circle::tag()->strokeLineJoin('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeMiterlimitValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingStrokeMiterlimitValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
@@ -610,7 +610,7 @@ final class CircleTest extends TestCase
         Circle::tag()->strokeMiterlimit('invalid-value');
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidStrokeOpacityValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingStrokeOpacityValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(