+++ tests/UsesTest.php
@@ -25,7 +25,7 @@ use UIAwesome\Html\Svg\Uses;
  * Test coverage.
  * - Accurate rendering of the `<use>` element.
  * - Correct application of global HTML attributes and SVG-specific attributes.
- * - Error handling for invalid attributes or configuration.
+ * - Error handling for invalid attribute values.
  * - Immutability of the API, ensuring that setting attributes returns a new instance.
  * - Integration with configuration providers and global factory defaults.
  * - Precedence of user-defined attributes over global defaults and provider settings.
@@ -374,7 +374,7 @@ final class UsesTest extends TestCase
         );
     }
 
-    public function testThrowInvalidArgumentExceptionForSettingInvalidOpacityValue(): void
+    public function testThrowInvalidArgumentExceptionForSettingOpacityValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(